requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
